Moving all server code to /src
