+++ src/server/server.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { renderToString } from 'react-dom/server';
-import Attendance from './client/attendance';
+import Attendance from '../client/attendance';
 
 const render = () => (
   renderToString(<Attendance />)
