+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/client.js',
+  entry: './src/client/index.js',
   output: {
     path: path.resolve(__dirname, 'assets'),
     filename: "client.js",
